+++ src/Toolkit/src/Kit/Lint/KitLinter.php
@@ -12,6 +12,7 @@
 namespace Symfony\UX\Toolkit\Kit\Lint;
 
 use Symfony\UX\Toolkit\Kit\Kit;
+use Symfony\UX\Toolkit\Kit\Lint\Checker\ClassMergeSpacingChecker;
 use Symfony\UX\Toolkit\Kit\Lint\Checker\ComponentDocChecker;
 use Symfony\UX\Toolkit\Kit\Lint\Checker\ComposerSymbolChecker;
 use Symfony\UX\Toolkit\Kit\Lint\Checker\CopyFilesExistenceChecker;
@@ -47,6 +48,7 @@ final class KitLinter
             new JsImportChecker(),
             new ComposerSymbolChecker(),
             new ComponentDocChecker(),
+            new ClassMergeSpacingChecker(),
         ];
     }